backend/requirements.txt
streamlit
fastapi
requests
uvicorn
aiohttp
google-genai
python-dotenv
pydantic
supabase
psycopg2-binary
python-multipart
sqlalchemy
python-jose
pymupdf
langgraph
langchain
beautifulsoup4
tiktoken
jsonschema
tenacity
cssselect